refacto: move utils in global includes
