+++ server/src/Network/TCP/TCPNetwork.cpp
@@ -6,11 +6,10 @@
 //
 
 #include <functional>
-#include <iostream>
 #include "TCPNetwork.hpp"
 #include "Network/Network.hpp"
 #include "libs/system/include/boost/system/detail/error_code.hpp"
-#include "logger.hpp"
+#include "utils/logger.hpp"
 
 network::TCPNetwork::TCPNetwork(const uint16_t port) :
     Network(port), _acceptor(this->_io_context), _endpoint(tcp::v4(), port)
